Wed, Dec 18, 2024, 2:41 PM -05:00

+++ bitcoin.py
@@ -2,9 +2,10 @@ import sys
 import requests
 
 try:
-    n = int(sys.argsv[1])
+    n = float(sys.argsv[1])
 except IndexError:
     sys.exit("Missing command line argument")
 except ValueError:
     sys.exit("Command-line argument is not a number")
+
